pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finlabair"
version = "0.1.0"
description = "A fun initiative cooked in the 1GA headquarters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.0",
    "openbb>=4.1.3",
    "pandas>=2.2.3",
    "yfinance>=0.2.51",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "matplotlib>=3.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
